+++ packages/next/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bidikit/next
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @bidikit/react@1.1.1
+
 ## 1.1.0
 
 ### Minor Changes